compact: relative output paths + much smaller pagination cursor
Output paths are now printed relative to the search root (read from the real
absolute path, but `sink_with_path` shows the stripped path), matching rg's
cwd-relative output and shrinking both the display and the cursor's keyset key.

Cursor shrink (assumes the caller pages from the same directory):
- Drop the absolute root from the cursor; store the original relative scope
  (None for the cwd) instead. This was the dominant size cost.
- Varint-encode the integer fields and length prefixes (small values -> 1 byte).
- Store the staleness fingerprint as its low 32 bits (advisory note only).
A typical cursor drops from ~170 to ~50 chars. Format version bumped to 2; old
cursors fail to decode (re-run the search), which is fine since they're ephemeral.

Threaded the search root through the confirm primitives so the daemon and the
in-process fallback both relativize uniformly. Removed nothing public; the unused
lib::search/stream_search helpers were kept (now take a root).

+++ src/compact.rs
@@ -77,16 +77,17 @@ impl Page {
             last_path: self.last_key.as_ref().map(|(p, _)| p.clone()),
             last_lineno: self.last_key.as_ref().map_or(0, |(_, l)| *l),
             prev_total: self.total_matches,
-            fingerprint: self.fingerprint,
+            fingerprint: self.fingerprint as u32,
             root_hint,
         })
     }
 
     /// A note for the caller to surface when resuming a cursor whose result set has since changed
-    /// (fingerprint mismatch), or `None`. `prev` is the cursor's `(prev_total, fingerprint)`.
-    pub fn staleness_note(&self, prev: Option<(usize, u64)>) -> Option<String> {
+    /// (fingerprint mismatch), or `None`. `prev` is the cursor's `(prev_total, fingerprint)`; the
+    /// fingerprint is the low 32 bits, so compare against this page's truncated to match.
+    pub fn staleness_note(&self, prev: Option<(usize, u32)>) -> Option<String> {
         match prev {
-            Some((prev_total, prev_fp)) if prev_fp != self.fingerprint => Some(format!(
+            Some((prev_total, prev_fp)) if prev_fp != self.fingerprint as u32 => Some(format!(
                 "result set changed since the previous page ({prev_total} -> {} matches)",
                 self.total_matches
             )),
